Move bench and specs to the tests folder
