+++ package.json
@@ -7,7 +7,7 @@
   "types": "index.d.ts",
   "author": "Dmitrii Goriunov",
   "scripts": {
-    "test": "mocha -r ts-node/register ./specs/*.spec.ts --exit",
+    "test": "mocha -r ts-node/register ./tests/specs/*.spec.ts --exit",
     "build": "clusterws-builder -type=node -npm",
     "publish": "npm run build && cd dist && npm publish && cd .."
   },
