+++ tests/benchmarks/emitter.bench.ts
